+++ internal/controllers/gateway/route_utils.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"reflect"
 
+	"github.com/go-logr/logr"
 	"github.com/samber/lo"
 	"github.com/samber/mo"
 	corev1 "k8s.io/api/core/v1"
@@ -15,7 +16,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
-	gatewayv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/gatewayapi"
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/util"
@@ -100,7 +100,7 @@ func parentRefsForRoute[T gatewayapi.RouteT](route T) ([]gatewayapi.ParentRefere
 		return nil, fmt.Errorf("can't determine parent Gateway for unsupported route type %s", reflect.TypeOf(route))
 	}
 	for _, ref := range refs {
-		if string(*ref.Group) != gatewayv1alpha2.GroupName || string(*ref.Kind) != "Gateway" {
+		if string(*ref.Group) != gatewayv1.GroupName || string(*ref.Kind) != "Gateway" {
 			return nil, fmt.Errorf("unsupported parent kind %s/%s", string(*ref.Group), string(*ref.Kind))
 		}
 	}
@@ -125,7 +125,7 @@ func parentRefsForRoute[T gatewayapi.RouteT](route T) ([]gatewayapi.ParentRefere
 // Gateway APIs route object (e.g. HTTPRoute, TCPRoute, e.t.c.) from the provided cached
 // client if they match this controller. If there are no gateways present for this route
 // OR the present gateways are references to missing objects, this will return a unsupportedGW error.
-func getSupportedGatewayForRoute[T gatewayapi.RouteT](ctx context.Context, mgrc client.Client, route T) ([]supportedGatewayWithCondition, error) {
+func getSupportedGatewayForRoute[T gatewayapi.RouteT](ctx context.Context, logger logr.Logger, mgrc client.Client, route T) ([]supportedGatewayWithCondition, error) {
 	// gather the parentrefs for this route object
 	parentRefs, err := parentRefsForRoute(route)
 	if err != nil {
@@ -158,6 +158,7 @@ func getSupportedGatewayForRoute[T gatewayapi.RouteT](ctx context.Context, mgrc 
 			}
 			return nil, fmt.Errorf("failed to retrieve gateway for route: %w", err)
 		}
+		gwLogger := logger.WithValues("parentRef.gateway", fmt.Sprintf("%s/%s", gateway.Namespace, gateway.Name))
 
 		// pull the GatewayClass for the Gateway object from the cached client
 		gatewayClass := gatewayapi.GatewayClass{}
@@ -195,15 +196,18 @@ func getSupportedGatewayForRoute[T gatewayapi.RouteT](ctx context.Context, mgrc 
 			allowedByAllowedRoutes  = false
 			allowedBySupportedKinds = false
 			allowedByListenerName   = false
+			listenerReady           = false
 		)
 
 		for _, listener := range gateway.Spec.Listeners {
+			listenerLogger := gwLogger.WithValues("listener", string(listener.Name))
 			// Check if the route matches listener's AllowedRoutes.
 			if ok, err := routeMatchesListenerAllowedRoutes(ctx, mgrc, route, listener, gateway.Namespace, parentRef.Namespace); err != nil {
 				return nil, fmt.Errorf("failed matching listener %s to a route %s for gateway %s: %w",
 					listener.Name, route.GetName(), gateway.Name, err,
 				)
 			} else if !ok {
+				listenerLogger.V(util.DebugLevel).Info("Route does not match listener's allowed routes")
 				continue
 			}
 			allowedByAllowedRoutes = true
@@ -211,16 +215,27 @@ func getSupportedGatewayForRoute[T gatewayapi.RouteT](ctx context.Context, mgrc 
 			// Check the listeners statuses:
 			// - Check if a listener status exists with a matching type (via SupportedKinds).
 			// - Check if it matches the requested listener by name (if specified).
-			// - And finally check if that listeners is marked as Ready.
-			if err := existsMatchingReadyListenerInStatus(route, listener, gateway.Status.Listeners); err != nil {
+			if err := existsMatchingListenerInStatus(route, listener, gateway.Status.Listeners); err != nil {
+				listenerLogger.V(util.DebugLevel).Info("Listener does not support this route", "reason", err.Error())
 				continue
 			} else { //nolint:revive
 				allowedBySupportedKinds = true
 			}
 
+			if err := listenerProgrammedInStatus(listener.Name, gateway.Status.Listeners); err != nil {
+				listenerLogger.V(util.DebugLevel).Info("Listener is not ready", "reason", err.Error())
+				continue
+			} else { //nolint:revive
+				listenerReady = true
+			}
+
 			// Check if listener name matches.
 			if parentRef.SectionName != nil {
 				if *parentRef.SectionName != "" && *parentRef.SectionName != listener.Name {
+					listenerLogger.V(util.DebugLevel).Info(
+						"Listener name does not match parentRef.SectionName",
+						"parentRef_sectionName", parentRef.SectionName,
+					)
 					continue
 				}
 				allowedByListenerName = true
@@ -231,12 +246,20 @@ func getSupportedGatewayForRoute[T gatewayapi.RouteT](ctx context.Context, mgrc 
 				if *parentRef.Port != listener.Port {
 					// This ParentRef has a port specified and it's different
 					// than current listener's port.
+					listenerLogger.V(util.DebugLevel).Info(
+						"Listener port does not match parentRef.Port",
+						"listener_port", listener.Port, "parentRef_port", parentRef.Port,
+					)
 					continue
 				}
 				portMatched = true
 			}
 
 			if !routeTypeMatchesListenerType(route, listener) {
+				listenerLogger.V(util.DebugLevel).Info(
+					"Route's type does not match listener's type",
+					"route_name", route.GetName(),
+				)
 				continue
 			}
 
@@ -246,6 +269,7 @@ func getSupportedGatewayForRoute[T gatewayapi.RouteT](ctx context.Context, mgrc 
 			} else {
 				condFalse := metav1.ConditionFalse
 				matchingHostname = &condFalse
+				listenerLogger.V(util.DebugLevel).Info("Route's hostname does not match listener's hostname")
 				continue
 			}
 
@@ -282,6 +306,8 @@ func getSupportedGatewayForRoute[T gatewayapi.RouteT](ctx context.Context, mgrc 
 				// If ParentRef specified listener names but none of the listeners matches the name,
 				// the gateway Status Condition Accepted must be set to False with reason RouteReasonNoMatchingParent.
 				reason = gatewayapi.RouteReasonNoMatchingParent
+			} else if !listenerReady {
+				reason = gatewayapi.RouteReasonNotAllowedByListeners
 			} else if (parentRef.Port != nil) && !portMatched {
 				// If ParentRef specified a Port but none of the listeners matched, the gateway Status
 				// Condition Accepted must be set to False with reason NoMatchingListenerPort
@@ -450,21 +476,21 @@ func routeMatchesListenerAllowedRoutes[T gatewayapi.RouteT](
 }
 
 var (
-	errUnsupportedRouteKind  = errors.New("unsupported route kind")
-	errUnmatchedListenerName = errors.New("unmatched listener name")
-	errListenerNotProgrammed = errors.New("no Programmed condition found for listener")
-	errListenerNotReadyYet   = errors.New("listener not ready yet")
+	errUnsupportedRouteKind          = errors.New("unsupported route kind")
+	errUnmatchedListenerName         = errors.New("unmatched listener name")
+	errListenerNoProgrammedCondition = errors.New("no Programmed condition found for listener")
+	errListenerNotProgrammedYet      = errors.New("listener not programmed yet")
 )
 
 // existsMatchingReadyListenerInStatus checks if:
 // - If a listener status exists with a matching type (via SupportedKinds).
 // - If it matches the requested listener by name (if specified).
 // - And finally check if the provided listener is marked as Ready.
-func existsMatchingReadyListenerInStatus[T gatewayapi.RouteT](route T, listener gatewayapi.Listener, lss []gatewayapi.ListenerStatus) error {
+func existsMatchingListenerInStatus[T gatewayapi.RouteT](route T, listener gatewayapi.Listener, lss []gatewayapi.ListenerStatus) error {
 	listenerFound := false
 
 	// Find listener's status...
-	listenerStatus, ok := lo.Find(lss, func(ls gatewayapi.ListenerStatus) bool {
+	_, ok := lo.Find(lss, func(ls gatewayapi.ListenerStatus) bool {
 		if ls.Name != listener.Name {
 			return false
 		}
@@ -500,15 +526,26 @@ func existsMatchingReadyListenerInStatus[T gatewayapi.RouteT](route T, listener 
 		return errUnsupportedRouteKind // Listener(s) found but none with matching supported kinds.
 	}
 
-	// ... and verify if it's programmed.
-	lReadyCond, ok := lo.Find(listenerStatus.Conditions, func(c metav1.Condition) bool {
-		return c.Type == string(gatewayapi.ListenerConditionProgrammed)
+	return nil
+}
+
+func listenerProgrammedInStatus(listenerName gatewayapi.SectionName, lss []gatewayapi.ListenerStatus) error {
+	listenerStatus, ok := lo.Find(lss, func(ls gatewayapi.ListenerStatus) bool {
+		return ls.Name == listenerName
 	})
 	if !ok {
-		return errListenerNotProgrammed
+		return errUnmatchedListenerName // Matching Listener's not found.
 	}
-	if lReadyCond.Status != "True" {
-		return errListenerNotReadyYet // Listener is not ready yet.
+
+	programmedStatus, ok := lo.Find(listenerStatus.Conditions, func(condition metav1.Condition) bool {
+		return condition.Type == string(gatewayapi.ListenerConditionProgrammed)
+	})
+	if !ok {
+		return errListenerNoProgrammedCondition // "Programmed" condition not found in conditions of listener's conditions.
+	}
+
+	if programmedStatus.Status != metav1.ConditionTrue {
+		return errListenerNotProgrammedYet // "Programmed" condition is not true.
 	}
 
 	return nil
@@ -905,8 +942,7 @@ func isRouteAcceptedByListener[T gatewayapi.RouteT](ctx context.Context,
 	// Check the listeners statuses:
 	// - Check if a listener status exists with a matching type (via SupportedKinds).
 	// - Check if it matches the requested listener by name (if specified).
-	// - And finally check if that listeners is marked as Ready.
-	if err := existsMatchingReadyListenerInStatus(route, listener, gateway.Status.Listeners); err != nil {
+	if err := existsMatchingListenerInStatus(route, listener, gateway.Status.Listeners); err != nil {
 		// return no error here, as we don't care of the reason why this check failed.
 		return false, nil //nolint:nilerr
 	}